+++ IntegrationBus/include/ib/sim/fr/IFrController.hpp
@@ -59,6 +59,9 @@ public:
     //! \brief Configure the controller and switch to ready state
     virtual void Configure(const ControllerConfig& config) = 0;
 
+    //! \brief Reconfigure a TX-Buffer that was previously setup with IFrController::Configure(const ControllerConfig&)
+    virtual void ReconfigureTxBuffer(uint16_t txBufferIdx, const TxBufferConfig& config) = 0;
+
     /*! \brief Update the content of a previously configured TX buffer.
      *
      * Due to the fixed and repetitive cycle of FlexRay, the behavior of UpdateTxBuffer is