+++ IntegrationBus/source/mw/comadapter/ComAdapter_impl.hpp
@@ -448,6 +448,12 @@ void ComAdapter<IbConnectionT>::SendIbMessage(EndpointAddress from, const sim::f
 }
 
 template <class IbConnectionT>
+void ComAdapter<IbConnectionT>::SendIbMessage(EndpointAddress from, const sim::fr::TxBufferConfigUpdate& msg)
+{
+    SendIbMessageImpl(from, msg);
+}
+
+template <class IbConnectionT>
 void ComAdapter<IbConnectionT>::SendIbMessage(EndpointAddress from, const sim::fr::TxBufferUpdate& msg)
 {
     SendIbMessageImpl(from, msg);