+++ IntegrationBus/source/mw/fastrtps/idl/FlexRayTopics.h
@@ -1977,6 +1977,200 @@ namespace ib
                     std::vector<ib::sim::fr::idl::TxBufferConfig> m_bufferConfigs;
                 };
                 /*!
+                 * @brief This class represents the structure TxBufferConfigUpdate defined by the user in the IDL file.
+                 * @ingroup FLEXRAYTOPICS
+                 */
+                class TxBufferConfigUpdate
+                {
+                public:
+
+                    /*!
+                     * @brief Default constructor.
+                     */
+                    eProsima_user_DllExport TxBufferConfigUpdate();
+
+                    /*!
+                     * @brief Default destructor.
+                     */
+                    eProsima_user_DllExport ~TxBufferConfigUpdate();
+
+                    /*!
+                     * @brief Copy constructor.
+                     * @param x Reference to the object ib::sim::fr::idl::TxBufferConfigUpdate that will be copied.
+                     */
+                    eProsima_user_DllExport TxBufferConfigUpdate(const TxBufferConfigUpdate &x);
+
+                    /*!
+                     * @brief Move constructor.
+                     * @param x Reference to the object ib::sim::fr::idl::TxBufferConfigUpdate that will be copied.
+                     */
+                    eProsima_user_DllExport TxBufferConfigUpdate(TxBufferConfigUpdate &&x);
+
+                    /*!
+                     * @brief Copy assignment.
+                     * @param x Reference to the object ib::sim::fr::idl::TxBufferConfigUpdate that will be copied.
+                     */
+                    eProsima_user_DllExport TxBufferConfigUpdate& operator=(const TxBufferConfigUpdate &x);
+
+                    /*!
+                     * @brief Move assignment.
+                     * @param x Reference to the object ib::sim::fr::idl::TxBufferConfigUpdate that will be copied.
+                     */
+                    eProsima_user_DllExport TxBufferConfigUpdate& operator=(TxBufferConfigUpdate &&x);
+
+                    /*!
+                     * @brief This function copies the value in member senderAddr
+                     * @param _senderAddr New value to be copied in member senderAddr
+                     */
+                    inline eProsima_user_DllExport void senderAddr(const ib::mw::idl::EndpointAddress &_senderAddr)
+                    {
+                        m_senderAddr = _senderAddr;
+                    }
+
+                    /*!
+                     * @brief This function moves the value in member senderAddr
+                     * @param _senderAddr New value to be moved in member senderAddr
+                     */
+                    inline eProsima_user_DllExport void senderAddr(ib::mw::idl::EndpointAddress &&_senderAddr)
+                    {
+                        m_senderAddr = std::move(_senderAddr);
+                    }
+
+                    /*!
+                     * @brief This function returns a constant reference to member senderAddr
+                     * @return Constant reference to member senderAddr
+                     */
+                    inline eProsima_user_DllExport const ib::mw::idl::EndpointAddress& senderAddr() const
+                    {
+                        return m_senderAddr;
+                    }
+
+                    /*!
+                     * @brief This function returns a reference to member senderAddr
+                     * @return Reference to member senderAddr
+                     */
+                    inline eProsima_user_DllExport ib::mw::idl::EndpointAddress& senderAddr()
+                    {
+                        return m_senderAddr;
+                    }
+                    /*!
+                     * @brief This function sets a value in member txBufferIndex
+                     * @param _txBufferIndex New value for member txBufferIndex
+                     */
+                    inline eProsima_user_DllExport void txBufferIndex(uint16_t _txBufferIndex)
+                    {
+                        m_txBufferIndex = _txBufferIndex;
+                    }
+
+                    /*!
+                     * @brief This function returns the value of member txBufferIndex
+                     * @return Value of member txBufferIndex
+                     */
+                    inline eProsima_user_DllExport uint16_t txBufferIndex() const
+                    {
+                        return m_txBufferIndex;
+                    }
+
+                    /*!
+                     * @brief This function returns a reference to member txBufferIndex
+                     * @return Reference to member txBufferIndex
+                     */
+                    inline eProsima_user_DllExport uint16_t& txBufferIndex()
+                    {
+                        return m_txBufferIndex;
+                    }
+                    /*!
+                     * @brief This function copies the value in member txBufferConfig
+                     * @param _txBufferConfig New value to be copied in member txBufferConfig
+                     */
+                    inline eProsima_user_DllExport void txBufferConfig(const ib::sim::fr::idl::TxBufferConfig &_txBufferConfig)
+                    {
+                        m_txBufferConfig = _txBufferConfig;
+                    }
+
+                    /*!
+                     * @brief This function moves the value in member txBufferConfig
+                     * @param _txBufferConfig New value to be moved in member txBufferConfig
+                     */
+                    inline eProsima_user_DllExport void txBufferConfig(ib::sim::fr::idl::TxBufferConfig &&_txBufferConfig)
+                    {
+                        m_txBufferConfig = std::move(_txBufferConfig);
+                    }
+
+                    /*!
+                     * @brief This function returns a constant reference to member txBufferConfig
+                     * @return Constant reference to member txBufferConfig
+                     */
+                    inline eProsima_user_DllExport const ib::sim::fr::idl::TxBufferConfig& txBufferConfig() const
+                    {
+                        return m_txBufferConfig;
+                    }
+
+                    /*!
+                     * @brief This function returns a reference to member txBufferConfig
+                     * @return Reference to member txBufferConfig
+                     */
+                    inline eProsima_user_DllExport ib::sim::fr::idl::TxBufferConfig& txBufferConfig()
+                    {
+                        return m_txBufferConfig;
+                    }
+
+                    /*!
+                     * @brief This function returns the maximum serialized size of an object
+                     * depending on the buffer alignment.
+                     * @param current_alignment Buffer alignment.
+                     * @return Maximum serialized size.
+                     */
+                    eProsima_user_DllExport static size_t getMaxCdrSerializedSize(size_t current_alignment = 0);
+
+                    /*!
+                     * @brief This function returns the serialized size of a data depending on the buffer alignment.
+                     * @param data Data which is calculated its serialized size.
+                     * @param current_alignment Buffer alignment.
+                     * @return Serialized size.
+                     */
+                    eProsima_user_DllExport static size_t getCdrSerializedSize(const ib::sim::fr::idl::TxBufferConfigUpdate& data, size_t current_alignment = 0);
+
+
+                    /*!
+                     * @brief This function serializes an object using CDR serialization.
+                     * @param cdr CDR serialization object.
+                     */
+                    eProsima_user_DllExport void serialize(eprosima::fastcdr::Cdr &cdr) const;
+
+                    /*!
+                     * @brief This function deserializes an object using CDR serialization.
+                     * @param cdr CDR serialization object.
+                     */
+                    eProsima_user_DllExport void deserialize(eprosima::fastcdr::Cdr &cdr);
+
+
+
+                    /*!
+                     * @brief This function returns the maximum serialized size of the Key of an object
+                     * depending on the buffer alignment.
+                     * @param current_alignment Buffer alignment.
+                     * @return Maximum serialized size.
+                     */
+                    eProsima_user_DllExport static size_t getKeyMaxCdrSerializedSize(size_t current_alignment = 0);
+
+                    /*!
+                     * @brief This function tells you if the Key has been defined for this type
+                     */
+                    eProsima_user_DllExport static bool isKeyDefined();
+
+                    /*!
+                     * @brief This function serializes the key members of an object using CDR serialization.
+                     * @param cdr CDR serialization object.
+                     */
+                    eProsima_user_DllExport void serializeKey(eprosima::fastcdr::Cdr &cdr) const;
+
+                private:
+                    ib::mw::idl::EndpointAddress m_senderAddr;
+                    uint16_t m_txBufferIndex;
+                    ib::sim::fr::idl::TxBufferConfig m_txBufferConfig;
+                };
+                /*!
                  * @brief This class represents the structure TxBufferUpdate defined by the user in the IDL file.
                  * @ingroup FLEXRAYTOPICS
                  */