+++ IntegrationBus/source/sim/fr/FrDatatypeUtils.hpp
@@ -15,6 +15,7 @@ bool operator==(const Frame& lhs, const Frame& rhs);
 bool operator==(const FrMessage& lhs, const FrMessage& rhs);
 bool operator==(const FrMessageAck& lhs, const FrMessageAck& rhs);
 bool operator==(const FrSymbol& lhs, const FrSymbol& rhs);
+bool operator==(const TxBufferConfigUpdate& lhs, const TxBufferConfigUpdate& rhs);
 bool operator==(const TxBufferUpdate& lhs, const TxBufferUpdate& rhs);
 bool operator==(const ControllerConfig& lhs, const ControllerConfig& rhs);
 bool operator==(const HostCommand& lhs, const HostCommand& rhs);